rename /start to /welcome
The name /welcome better conveys that it is for first-time users, not the page where you start each session.
