+++ web/src/enterprise/routes.tsx
@@ -1,5 +1,18 @@
 import React from 'react'
+import { Redirect } from 'react-router'
 import { LayoutRouteProps, routes } from '../routes'
+const WelcomeMainPage = React.lazy(async () => ({
+    default: (await import('./dotcom/welcome/WelcomeMainPage')).WelcomeMainPage,
+}))
+const WelcomeSearchPage = React.lazy(async () => ({
+    default: (await import('./dotcom/welcome/WelcomeSearchPage')).WelcomeSearchPage,
+}))
+const WelcomeCodeIntelligencePage = React.lazy(async () => ({
+    default: (await import('./dotcom/welcome/WelcomeCodeIntelligencePage')).WelcomeCodeIntelligencePage,
+}))
+const WelcomeIntegrationsPage = React.lazy(async () => ({
+    default: (await import('./dotcom/welcome/WelcomeIntegrationsPage')).WelcomeIntegrationsPage,
+}))
 const NewProductSubscriptionPageOrRedirectUser = React.lazy(async () => ({
     default: (await import('./user/productSubscriptions/NewProductSubscriptionPageOrRedirectUser'))
         .NewProductSubscriptionPageOrRedirectUser,
@@ -12,6 +25,31 @@ export const enterpriseRoutes: ReadonlyArray<LayoutRouteProps> = [
         path: '/user/subscriptions/new',
         exact: true,
         render: props => <NewProductSubscriptionPageOrRedirectUser {...props} />,
+    },
+    {
+        path: '/start',
+        render: () => <Redirect to="/welcome" />,
+        exact: true,
+    },
+    {
+        path: '/welcome',
+        render: props => <WelcomeMainPage {...props} />,
+        exact: true,
+    },
+    {
+        path: '/welcome/search',
+        render: props => <WelcomeSearchPage {...props} />,
+        exact: true,
+    },
+    {
+        path: '/welcome/code-intelligence',
+        render: props => <WelcomeCodeIntelligencePage {...props} />,
+        exact: true,
+    },
+    {
+        path: '/welcome/integrations',
+        render: props => <WelcomeIntegrationsPage {...props} />,
+        exact: true,
     },
     ...routes,
 ]
